Crud adicionar adiministrador, update static e templates

+++ src/main/java/br/com/hotel/hotel/repositories/AdministradorRepositories.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.hotel.hotel.entities.Administrador;
 
-public interface AdministradorRepositories extends JpaRepository<Administrador, Integer>{
+public interface AdministradorRepositories extends JpaRepository<Administrador, Long>{
     
 }
